lmsBackend/Accounts/serializers.py
from rest_framework import serializers
from .models import User, Messages


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Messages
        fields = '__all__' 

class UserModelSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['id','first_name','last_name','contact','gender','email']
