requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.33.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
